자료형 int로 통합

+++ doonut-support/src/main/java/com/doonutmate/image/ImageMeta.java
@@ -1,4 +1,4 @@
 package com.doonutmate.image;
 
-public record ImageMeta(int widht, int height, long capacity) {
+public record ImageMeta(int widht, int height, int capacity) {
 }
